requirements.txt
# Core Django
Django==4.2.7
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
django-cors-headers==4.3.1
django-filter==23.3

# Image processing
Pillow>=9.0.0

# WebSocket and real-time
channels==4.0.0
channels-redis==4.1.0
daphne==4.0.0
redis==5.0.1

# Background tasks
celery==5.3.4
django-celery-beat==2.5.0
django-celery-results==2.6.0

# Environment and configuration
python-decouple==3.8

# Development tools
django-extensions==3.2.3

# System monitoring
psutil==5.9.6

# Database (use SQLite for development, PostgreSQL for production)
psycopg2-binary==2.9.9

# File storage (optional for development)
django-storages==1.14.2
boto3==1.34.0

# Production server
gunicorn==21.2.0
whitenoise==6.6.0

# Development and testing tools
flake8==6.1.0
black==23.11.0
isort==5.12.0
bandit==1.7.5
safety==2.3.5